add api module (#29)
