+++ stac_api/app.py
@@ -1,6 +1,6 @@
 """FastAPI application."""
+from stac_api.api import create_app
 from stac_api.config import ApiSettings
-from stac_api.create_app import create_app
 
 settings = ApiSettings(
     stac_api_extensions=["context", "fields", "query", "sort", "transaction"],
